Reorganizing api calls into separate .js files
