+++ backend/src/commands/features.js
@@ -1,0 +1,12 @@
+import AWS from "aws-sdk";
+import dotenv from "dotenv";
+dotenv.config();
+
+AWS.config.update({
+  region: process.env.AWS_DEFAULT_REGION,
+  accesKeyId: process.env.AWS_ACCESS_KEY_ID,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+});
+
+export const DocumentClient = new AWS.DynamoDB.DocumentClient();
+export const TABLE_NAME = "features";
