+++ backend/src/commands/properties.js
@@ -1,0 +1,52 @@
+import AWS from "aws-sdk";
+import dotenv from "dotenv";
+dotenv.config();
+
+AWS.config.update({
+  region: process.env.AWS_DEFAULT_REGION,
+  accesKeyId: process.env.AWS_ACCESS_KEY_ID,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+});
+
+export const DocumentClient = new AWS.DynamoDB.DocumentClient();
+export const TABLE_NAME = "properties";
+
+
+export const getSinglePropertyByID = async (TABLE_NAME, id) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Key: {
+      id,
+    },
+  };
+  const item = await DocumentClient.get(params).promise();
+  console.log(item);
+  return item;
+};
+
+export const getAllProperties = async () => {
+  const params = {
+    TableName: TABLE_NAME,
+  };
+  const items = await DocumentClient.scan(params).promise();
+  console.log(items);
+  return items;
+};
+
+export const insertProperty = async (TABLE_NAME, itemObject) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Item: itemObject,
+  };
+  return await DocumentClient.put(params).promise();
+};
+
+export const deleteSinglePropertyById = async (TABLE_NAME, id) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Key: {
+      id,
+    },
+  };
+  return await DocumentClient.delete(params).promise();
+};
