+++ backend/src/commands/tasks.js
@@ -1,0 +1,41 @@
+import AWS from "aws-sdk";
+import dotenv from "dotenv";
+dotenv.config();
+
+AWS.config.update({
+  region: process.env.AWS_DEFAULT_REGION,
+  accesKeyId: process.env.AWS_ACCESS_KEY_ID,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+});
+
+export const DocumentClient = new AWS.DynamoDB.DocumentClient();
+export const TABLE_NAME = "tasks";
+
+
+export const getAllTasks = async () => {
+  const params = {
+    TableName: TABLE_NAME,
+  };
+  const items = await DocumentClient.scan(params).promise();
+  console.log(items);
+  return items;
+};
+
+
+export const insertTask = async (itemObject) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Item: itemObject,
+  };
+  return await DocumentClient.put(params).promise();
+};
+
+export const deleteSingleTaskById = async (TABLE_NAME, id) => {
+  const params = {
+    TableName: TABLE_NAME,
+    Key: {
+      id,
+    },
+  };
+  return await DocumentClient.delete(params).promise();
+};
